fix: delete controller manager
