--- packages/design/src/physics/characterkinematic/index.ts
@@ -1,2 +1,0 @@
-export type { ICharacterControllerManager } from "./ICharacterControllerManager";
-export type { ICharacterController } from "./ICharacterController";
