+++ packages/design/src/physics/index.ts
@@ -4,5 +4,5 @@ export type { ICollider } from "./ICollider";
 export type { IStaticCollider } from "./IStaticCollider";
 export type { IDynamicCollider } from "./IDynamicCollider";
 export type { IPhysicsMaterial } from "./IPhysicsMaterial";
+export type { ICharacterController } from "./ICharacterController";
 export * from "./shape/index";
-export * from "./characterkinematic/index";
